Consolidated SLOC functionality, returning stats for all languages.
SLOC functionality is now consolidated in a single module instead of
being in two separate modules, with some of the functionality
duplicated between the two modules.

Also, line and file counts are now returned for all languages, not
just for PHP. This is returned as a map that can then be filtered to
just return PHP-related SLOC info. Keyword params on existing routines
to get the counts enable this filter while also allowing counts to
be retrieved for other languages.