Add Closure support for CFG generator
Closure support added to generation of control flow graphs. Each
closure is extracted as an individual function with an anonymous
name. The closure is not linked into the CFG of the surrounding
unit (function, script, method) except as a single statement,
similarly to how functions contained in scripts are handled.